fix: do not use Mix during runtime
Mix is not bundled with OTP releases, so we cannot rely on it to compute
the migrations directory during runtime. This commit creates a new
function for runtime usage that mimics the logic of the Mix.Ecto
implementation, but uses the application build dir as a base instead of
the mix project dir.